Fixed ordercontroller update test

+++ src/main/java/com/qa/ims/persistence/dao/OrderDAO.java
@@ -39,11 +39,12 @@ public class OrderDAO implements Dao<Order> {
 	public List<Order> readAll() {
 		try (Connection connection = DBUtils.getInstance().getConnection();
 				Statement statement = connection.createStatement();
-				ResultSet resultSet = statement.executeQuery("SELECT order_items.id, order_items.order_id, orders.customer_id, "
-						+ "customers.first_name, customers.surname " + "FROM ((order_items "
-						+ "JOIN orders ON order_items.order_id = orders.id)" 
-						+ " JOIN customers ON orders.customer_id = customers.id)"
-						+ "GROUP BY order_items.order_id;");) {
+				ResultSet resultSet = statement
+						.executeQuery("SELECT order_items.id, order_items.order_id, orders.customer_id, "
+								+ "customers.first_name, customers.surname " + "FROM ((order_items "
+								+ "JOIN orders ON order_items.order_id = orders.id)"
+								+ " JOIN customers ON orders.customer_id = customers.id)"
+								+ "GROUP BY order_items.order_id;");) {
 			List<Order> orders = new ArrayList<>();
 			while (resultSet.next()) {
 				orders.add(modelFromResultSet(resultSet));
@@ -59,10 +60,13 @@ public class OrderDAO implements Dao<Order> {
 	public Order readLatest() {
 		try (Connection connection = DBUtils.getInstance().getConnection();
 				Statement statement = connection.createStatement();
-				ResultSet resultSet = statement.executeQuery("SELECT order_items.id, order_items.order_id, orders.customer_id, "
-						+ "customers.first_name, customers.surname, order_items.item_id, items.item_name, items.price " + "FROM (((order_items"
-						+ " JOIN orders ON order_items.order_id = orders.id)" + " JOIN customers ON orders.customer_id = customers.id)"
-						+ " JOIN items ON order_items.item_id = items.id)" + " ORDER BY order_items.id DESC LIMIT 1;");) {
+				ResultSet resultSet = statement
+						.executeQuery("SELECT order_items.id, order_items.order_id, orders.customer_id, "
+								+ "customers.first_name, customers.surname, order_items.item_id, items.item_name, items.price "
+								+ "FROM (((order_items" + " JOIN orders ON order_items.order_id = orders.id)"
+								+ " JOIN customers ON orders.customer_id = customers.id)"
+								+ " JOIN items ON order_items.item_id = items.id)"
+								+ " ORDER BY order_items.id DESC LIMIT 1;");) {
 			resultSet.next();
 			return modelFromResultSet(resultSet);
 		} catch (Exception e) {
@@ -71,14 +75,17 @@ public class OrderDAO implements Dao<Order> {
 		}
 		return null;
 	}
-	
+
 	@Override
 	public Order read(Long id) {
 		try (Connection connection = DBUtils.getInstance().getConnection();
-				PreparedStatement statement = connection.prepareStatement("SELECT order_items.id, order_items.order_id, orders.customer_id, "
-						+ "customers.first_name, customers.surname, order_items.item_id, items.item_name, items.price " + "FROM (((order_items "
-						+ "JOIN orders ON order_items.order_id = orders.id)" + " JOIN customers  ON orders.customer_id = customers.id)"
-						+ " JOIN items on order_items.item_id = items.id)" + "WHERE order_items.order_id = ?;");) {
+				PreparedStatement statement = connection
+						.prepareStatement("SELECT order_items.id, order_items.order_id, orders.customer_id, "
+								+ "customers.first_name, customers.surname, order_items.item_id, items.item_name, items.price "
+								+ "FROM (((order_items " + "JOIN orders ON order_items.order_id = orders.id)"
+								+ " JOIN customers  ON orders.customer_id = customers.id)"
+								+ " JOIN items on order_items.item_id = items.id)"
+								+ "WHERE order_items.order_id = ?;");) {
 			statement.setLong(1, id);
 			try (ResultSet resultSet = statement.executeQuery();) {
 				resultSet.next();
@@ -93,10 +100,10 @@ public class OrderDAO implements Dao<Order> {
 
 	private List<Item> readItemsInOrder(Long orderId) {
 		try (Connection connection = DBUtils.getInstance().getConnection();
-				PreparedStatement statement = connection.prepareStatement("SELECT items.id, items.item_name, items.price "
-						+ "FROM order_items"
-						+ " JOIN items ON order_items.item_id = items.id"
-						+ " WHERE order_items.order_id = ?;");) {
+				PreparedStatement statement = connection
+						.prepareStatement("SELECT items.id, items.item_name, items.price " + "FROM order_items"
+								+ " JOIN items ON order_items.item_id = items.id"
+								+ " WHERE order_items.order_id = ?;");) {
 			statement.setLong(1, orderId);
 			List<Item> items = new ArrayList<Item>();
 			List<String> itemStrings = new ArrayList<String>();
@@ -121,26 +128,36 @@ public class OrderDAO implements Dao<Order> {
 	 */
 	@Override
 	public Order create(Order order) {
-		if (!customerHasCurrentOrder(order.getCustomerId())) {
-			try (Connection connection = DBUtils.getInstance().getConnection();
-					PreparedStatement statement1 = connection
-							.prepareStatement("INSERT INTO orders(customer_id) VALUES (?)");
-					PreparedStatement statement2 = connection
-							.prepareStatement("INSERT INTO order_items(order_id, item_id) VALUES (?, ?)");) {
-				statement1.setLong(1, order.getCustomerId());
-				statement1.executeUpdate();
-				statement2.setLong(1, getOrderId(order.getCustomerId()));
-				statement2.setLong(2, order.getItems().get(0).getId());
-				statement2.executeUpdate();
-				return readLatest();
-			} catch (Exception e) {
-				LOGGER.debug(e);
-				LOGGER.error(e.getMessage());
+		if (customerExists(order.getCustomerId())) {
+			if (itemExists(order.getItemId())) {
+				if (!customerHasCurrentOrder(order.getCustomerId())) {
+					try (Connection connection = DBUtils.getInstance().getConnection();
+							PreparedStatement statement1 = connection
+									.prepareStatement("INSERT INTO orders(customer_id) VALUES (?)");
+							PreparedStatement statement2 = connection
+									.prepareStatement("INSERT INTO order_items(order_id, item_id) VALUES (?, ?)");) {
+						statement1.setLong(1, order.getCustomerId());
+						statement1.executeUpdate();
+						statement2.setLong(1, getOrderId(order.getCustomerId()));
+						statement2.setLong(2, order.getItems().get(0).getId());
+						statement2.executeUpdate();
+						return readLatest();
+					} catch (Exception e) {
+						LOGGER.debug(e);
+						LOGGER.error(e.getMessage());
+					}
+					return null;
+				} else {
+					LOGGER.info("CUSTOMER CANNOT HAVE TWO ORDERS AT THE SAME TIME."
+							+ " CLOSE PREVIOUS ORDERS TO CREATE A NEW ONE OR UPDATE THE PREVIOUS ORDER WITH NEW ITEMS");
+					return null;
+				}
+			} else {
+				LOGGER.info("ITEM DOES NOT EXIST. PLEASE CHOOSE AN EXISTING ITEM.");
+				return null;
 			}
-			return null;
 		} else {
-			LOGGER.info("CUSTOMER CANNOT HAVE TWO ORDERS AT THE SAME TIME."
-					+ " CLOSE PREVIOUS ORDERS TO CREATE A NEW ONE OR UPDATE THE PREVIOUS ORDER WITH NEW ITEMS");
+			LOGGER.info("CUSTOMER DOES NOT EXIST. PLEASE CHOOSE AN EXISTING CUSTOMER.");
 			return null;
 		}
 	}
@@ -160,11 +177,10 @@ public class OrderDAO implements Dao<Order> {
 		}
 		return null;
 	}
-	
+
 	public boolean customerExists(Long customerId) {
 		try (Connection connection = DBUtils.getInstance().getConnection();
-				PreparedStatement statement = connection
-						.prepareStatement("SELECT * FROM customers WHERE id = ?");) {
+				PreparedStatement statement = connection.prepareStatement("SELECT * FROM customers WHERE id = ?");) {
 			statement.setLong(1, customerId);
 			try (ResultSet resultSet = statement.executeQuery();) {
 				if (resultSet.next()) {
@@ -179,11 +195,10 @@ public class OrderDAO implements Dao<Order> {
 		}
 		return false;
 	}
-	
+
 	public boolean itemExists(Long itemId) {
 		try (Connection connection = DBUtils.getInstance().getConnection();
-				PreparedStatement statement = connection
-						.prepareStatement("SELECT * FROM items WHERE id = ?");) {
+				PreparedStatement statement = connection.prepareStatement("SELECT * FROM items WHERE id = ?");) {
 			statement.setLong(1, itemId);
 			try (ResultSet resultSet = statement.executeQuery();) {
 				if (resultSet.next()) {
@@ -198,11 +213,10 @@ public class OrderDAO implements Dao<Order> {
 		}
 		return false;
 	}
-	
+
 	public boolean orderExists(Long orderId) {
 		try (Connection connection = DBUtils.getInstance().getConnection();
-				PreparedStatement statement = connection
-						.prepareStatement("SELECT * FROM orders WHERE id = ?");) {
+				PreparedStatement statement = connection.prepareStatement("SELECT * FROM orders WHERE id = ?");) {
 			statement.setLong(1, orderId);
 			try (ResultSet resultSet = statement.executeQuery();) {
 				if (resultSet.next()) {
@@ -237,8 +251,6 @@ public class OrderDAO implements Dao<Order> {
 		return false;
 	}
 
-	
-
 	/**
 	 * Updates an order in the database
 	 * 
@@ -259,26 +271,37 @@ public class OrderDAO implements Dao<Order> {
 	}
 
 	public Order removeItem(Long id, Long itemId) {
-		if (orderContainsItem(id, itemId)) {
-		try (Connection connection = DBUtils.getInstance().getConnection();
-				PreparedStatement statement = connection
-						.prepareStatement("DELETE FROM order_items WHERE order_id = ? AND item_id = ?;");) {
-			statement.setLong(1, id);
-			statement.setLong(2, itemId);
-			statement.executeUpdate();
-		if (orderHasItems(id)) {
-			return read(id);
+		if (orderExists(id)) {
+			if (itemExists(itemId)) {
+				if (orderContainsItem(id, itemId)) {
+					try (Connection connection = DBUtils.getInstance().getConnection();
+							PreparedStatement statement = connection
+									.prepareStatement("DELETE FROM order_items WHERE order_id = ? AND item_id = ?;");) {
+						statement.setLong(1, id);
+						statement.setLong(2, itemId);
+						statement.executeUpdate();
+						if (orderHasItems(id)) {
+							return read(id);
+						} else {
+							delete(id);
+							return null;
+						}
+					} catch (Exception e) {
+						LOGGER.debug(e);
+						LOGGER.error(e.getMessage());
+					}
+					return null;
+				} else {
+					LOGGER.info(
+							"ORDER DOES NOT CONTAIN SELECTED ITEM. YOU CAN ONLY REMOVE AN ITEM THAT THE ORDER CONTAINS");
+					return null;
+				}
 			} else {
-				delete(id);
+				LOGGER.info("ITEM DEOS NOT EXIST. PLEASE ENTER AN EXISTING ITEM");
 				return null;
-			} 
-		} catch (Exception e) {
-			LOGGER.debug(e);
-			LOGGER.error(e.getMessage());
-		}
-		return null;
+			}
 		} else {
-			LOGGER.info("ORDER DOES NOT CONTAIN SELECTED ITEM. YOU CAN ONLY REMOVE AN ITEM THAT THE ORDER CONTAINS");
+			LOGGER.info("ORDER DOES NOT EXIST PLEASE ENTER AN EXISTING ORDER");
 			return null;
 		}
 	}
@@ -297,8 +320,7 @@ public class OrderDAO implements Dao<Order> {
 		}
 		return null;
 	}
-	
-	
+
 	public boolean orderHasItems(Long orderId) {
 		try (Connection connection = DBUtils.getInstance().getConnection();
 				PreparedStatement statement = connection
@@ -317,7 +339,7 @@ public class OrderDAO implements Dao<Order> {
 		}
 		return false;
 	}
-	
+
 	private boolean orderContainsItem(Long orderId, Long itemId) {
 		try (Connection connection = DBUtils.getInstance().getConnection();
 				PreparedStatement statement = connection
@@ -346,11 +368,12 @@ public class OrderDAO implements Dao<Order> {
 	@Override
 	public int delete(long id) {
 		try (Connection connection = DBUtils.getInstance().getConnection();
-				PreparedStatement statement1 = connection.prepareStatement("DELETE FROM order_items WHERE order_id = ?");
+				PreparedStatement statement1 = connection
+						.prepareStatement("DELETE FROM order_items WHERE order_id = ?");
 				PreparedStatement statement2 = connection.prepareStatement("DELETE FROM orders WHERE id = ?");) {
 			statement1.setLong(1, id);
 			statement2.setLong(1, id);
-			statement1.executeUpdate();	
+			statement1.executeUpdate();
 			return statement2.executeUpdate();
 		} catch (Exception e) {
 			LOGGER.debug(e);
